Fixed GenericForeignKey errors with shadow models, move user statictics objects out of hub app
