+++ kobo/settings/base.py
@@ -98,7 +98,8 @@ INSTALLED_APPS = (
     'kobo.apps.external_integrations.ExternalIntegrationsAppConfig',
     'markdownx',
     'kobo.apps.help',
-    'kobo.apps.user_statistics'
+    'kobo.apps.shadow_model.ShadowModelAppConfig',
+    'kobo.apps.user_statistics.UserStatisticsAppConfig',
 )
 
 MIDDLEWARE = [
